requirements.txt
pytorch-lightning
wandb
sentence-transformers
simple_parsing
torch_sparse
torch_geometric